Vue в index.js лишнее убрал. Корзина идеальная
