--- Students/Vyacheslav Ovchinnikov/Project/src/components/basket.js
@@ -1,45 +1,0 @@
-import List from './list.js';
-
-export default class Basket extends List {
-    constructor(container = '#basket', url = '/basket.json') {
-        super(container, url);
-        this.containerItems = document.querySelector('#basket-items');
-        this.shown = false;
-    }
-
-    _handleActions() {
-        document.querySelector('#basket-toggler').addEventListener('click', () => {
-            this.container.classList.toggle('invisible');
-            // document.querySelector('#basket').classList.toggle('invisible');
-            this.shown = !this.shown;
-        })
-
-        this.container.addEventListener('click', ev => {
-            if (ev.target.name == 'remove') {
-                this._remove(ev.target.dataset.id);
-            }
-        })
-    }
-
-    add(item) {
-        let find = this.items.find(el => el.productId == item.productId);
-        if (find) {
-            find.amount++;
-        } else {
-            let newItem = Object.assign({}, item, { amount: 1 });
-            console.log(newItem)
-            this.items.push(newItem);
-        }
-        this._render();
-    }
-
-    _remove(id) {
-        let find = this.items.find(el => el.productId == id);
-        if (find.amount > 1) {
-            find.amount--;
-        } else {
-            this.items.splice(this.items.indexOf(find), 1);
-        }
-        this._render();
-    }
-}
